test(Grid): improve test grid constructor function
Improve the internal test function grid_with_occupied_corners_and_origin by making it handle cases size == 0 and size == 1. The function was never used with size == 1 previously but it is a useful test case. If size == 0 the function panics (as it should) because this is an invalid test; a grid of size zero cannot have an occupied origin.